Move packages from internal to common
