+++ common/deadlock/deadlock.go
@@ -32,11 +32,11 @@ import (
 
 	"go.temporal.io/server/common/clock"
 	"go.temporal.io/server/common/dynamicconfig"
+	"go.temporal.io/server/common/goro"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/pingable"
-	"go.temporal.io/server/internal/goro"
 	"go.uber.org/fx"
 	"google.golang.org/grpc/health"
 )
