+++ common/dynamicconfig/collection.go
@@ -37,12 +37,12 @@ import (
 
 	"github.com/mitchellh/mapstructure"
 	"go.temporal.io/server/common/clock"
+	"go.temporal.io/server/common/goro"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/pingable"
 	"go.temporal.io/server/common/primitives/timestamp"
 	"go.temporal.io/server/common/util"
-	"go.temporal.io/server/internal/goro"
 	"google.golang.org/protobuf/reflect/protoreflect"
 )
