+++ common/effect/buffer_test.go
@@ -29,7 +29,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"go.temporal.io/server/internal/effect"
+	"go.temporal.io/server/common/effect"
 )
 
 func TestBufferApplyOrder(t *testing.T) {
