+++ common/effect/immediate.go
