+++ common/finalizer/finalizer.go
@@ -31,10 +31,10 @@ import (
 
 	"github.com/pkg/errors"
 	cclock "go.temporal.io/server/common/clock"
+	"go.temporal.io/server/common/goro"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
-	"go.temporal.io/server/internal/goro"
 )
 
 var (
