+++ common/finalizer/finalizer_test.go
@@ -32,10 +32,10 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.temporal.io/server/common/clock"
 	. "go.temporal.io/server/common/finalizer"
+	"go.temporal.io/server/common/goro"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/metrics/metricstest"
-	"go.temporal.io/server/internal/goro"
 )
 
 func TestFinalizer(t *testing.T) {
