+++ common/goro/goro_test.go
@@ -32,7 +32,7 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"go.temporal.io/server/internal/goro"
+	"go.temporal.io/server/common/goro"
 )
 
 func blockOnCtxReturnErr(ctx context.Context) error {
