+++ common/goro/group_test.go
@@ -22,22 +22,47 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package effect_test
+package goro_test
 
 import (
 	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"go.temporal.io/server/internal/effect"
+	"go.temporal.io/server/common/goro"
 )
 
-func TestImmediate(t *testing.T) {
-	var i int
-	immediate := effect.Immediate(context.TODO())
-	immediate.OnAfterCommit(func(context.Context) { i = 1 })
-	require.Equal(t, i, 1, "commit func should have run")
+func TestMultiCancelAndWait(t *testing.T) {
+	var g goro.Group
+	g.Go(blockOnCtxReturnNil)
+	g.Go(blockOnCtxReturnNil)
+	g.Go(blockOnCtxReturnNil)
+	g.Cancel()
+	g.Wait()
+}
 
-	immediate.OnAfterRollback(func(context.Context) { i = 2 })
-	require.Equal(t, i, 1, "rollback func should not run")
+func TestCancelBeforeGo(t *testing.T) {
+	var g goro.Group
+	g.Cancel()
+	g.Go(func(ctx context.Context) error {
+		require.ErrorIs(t, context.Canceled, ctx.Err())
+		return nil
+	})
+	g.Wait()
+}
+
+func TestWaitOnNothing(t *testing.T) {
+	var g goro.Group
+	g.Wait() // nothing running, should return immediately
+}
+
+func TestWaitOnDifferentThread(t *testing.T) {
+	var g goro.Group
+	g.Go(blockOnCtxReturnNil)
+	h := goro.NewHandle(context.TODO()).Go(func(context.Context) error {
+		g.Wait()
+		return nil
+	})
+	g.Cancel()
+	<-h.Done()
 }
