+++ common/nexus/nexustest/server.go
@@ -32,7 +32,7 @@ import (
 
 	"github.com/nexus-rpc/sdk-go/nexus"
 	"github.com/stretchr/testify/require"
-	"go.temporal.io/server/internal/freeport"
+	"go.temporal.io/server/common/freeport"
 )
 
 func AllocListenAddress() string {
