+++ common/protocol/naming_test.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 	protocolpb "go.temporal.io/api/protocol/v1"
-	"go.temporal.io/server/internal/protocol"
+	"go.temporal.io/server/common/protocol"
 	"google.golang.org/protobuf/types/known/anypb"
 	"google.golang.org/protobuf/types/known/emptypb"
 )
