+++ service/history/shard/ownership.go
@@ -29,12 +29,12 @@ import (
 	"time"
 
 	"go.temporal.io/server/common/convert"
+	"go.temporal.io/server/common/goro"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/membership"
 	"go.temporal.io/server/common/metrics"
 	serviceerrors "go.temporal.io/server/common/serviceerror"
-	"go.temporal.io/server/internal/goro"
 	"go.temporal.io/server/service/history/configs"
 )
