+++ service/history/workflow/update/registry_test.go
@@ -36,9 +36,9 @@ import (
 	"go.temporal.io/api/serviceerror"
 	updatepb "go.temporal.io/api/update/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
+	"go.temporal.io/server/common/effect"
 	. "go.temporal.io/server/common/testing/protoutils"
 	"go.temporal.io/server/common/testing/testvars"
-	"go.temporal.io/server/internal/effect"
 	"go.temporal.io/server/service/history/consts"
 	"go.temporal.io/server/service/history/workflow/update"
 	"google.golang.org/protobuf/types/known/anypb"
