+++ service/history/workflow/update/store.go
@@ -30,7 +30,7 @@ import (
 	historypb "go.temporal.io/api/history/v1"
 	updatepb "go.temporal.io/api/update/v1"
 	persistencespb "go.temporal.io/server/api/persistence/v1"
-	"go.temporal.io/server/internal/effect"
+	"go.temporal.io/server/common/effect"
 )
 
 type (
